crates/sel4-initialize-tls: Add on-heap feature
