+++ crates/sel4-initialize-tls/Cargo.nix
@@ -13,5 +13,7 @@ mk {
   };
   features = {
     on-stack = [];
+    on-heap = [ "alloc" ];
+    alloc = [];
   };
 }
